changed exported properties of licensed asset database to storage only
This hides the properties in the editor, thus disabling direct editing of the
database through the inspector.
For now, this is done to avoid modification outside of the provided
interface and therefore potential state disrepancies if the interface
does not react to certain changes for some reason.
The change may be reverted in the future if the UI is sufficiently
robust.